20064 【未排程结果】未排的都是，全部没有排，应该也包括部分未排

+++ APS-Modules/aps-factory-engine/src/main/java/com/zlt/aps/factory/utils/NoProductionPlanUtils.java
@@ -45,7 +45,7 @@ public class NoProductionPlanUtils {
         return;
       }
       // 即没有排产计划，又不是不排产计划
-      if (unProductionQty > BigDecimal.ZERO.longValue()) {
+      if (unProductionQty > BigDecimal.ZERO.intValue()) {
         noProductionPlanList.add(noProductionPlan);
       }
     });
